+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneAppProtectionPolicyAndroid/MSFT_IntuneAppProtectionPolicyAndroid.psm1
@@ -245,7 +245,7 @@ function Get-TargetResource
             {
                 Write-Verbose -Message "Searching for Policy using Id {$Id}"
                 $policyInfo = Get-MgBetaDeviceAppManagementAndroidManagedAppProtection -Filter "Id eq '$Id'" -ExpandProperty Apps, assignments `
-                    -ErrorAction Stop
+                    -ErrorAction SilentlyContinue
             }
 
             if ($null -eq $policyInfo)
@@ -296,16 +296,24 @@ function Get-TargetResource
         {
             foreach ($assignment in $policyInfo.Assignments)
             {
+                $groupInfo = Get-MgGroup -GroupId $assignment.Target.AdditionalProperties.groupId -ErrorAction SilentlyContinue
+                $groupValue = $assignment.Target.AdditionalProperties.groupId
+                if ($null -ne $groupInfo)
+                {
+                    $groupValue = $groupInfo.DisplayName
+                }
+
                 switch ($assignment.Target.AdditionalProperties.'@odata.type')
                 {
                     '#microsoft.graph.groupAssignmentTarget'
                     {
-                        $assignmentsArray += $assignment.Target.AdditionalProperties.groupId
+
+                        $assignmentsArray += $groupValue
                     }
 
                     '#microsoft.graph.exclusionGroupAssignmentTarget'
                     {
-                        $exclusionArray += $assignment.Target.AdditionalProperties.groupId
+                        $exclusionArray += $groupValue
                     }
                 }
             }
@@ -361,6 +369,7 @@ function Get-TargetResource
     }
     catch
     {
+        Write-Verbose -Message $_
         if ($_.Exception.Message -eq 'Multiple Policies with same displayname identified - Module currently only functions with unique names')
         {
             throw $_
@@ -662,7 +671,13 @@ function Set-TargetResource
         $PSBoundParameters.Assignments | ForEach-Object {
             if ($_ -ne $null)
             {
-                $assignmentsArray += set-JSONstring -id $_ -type 'Assignments'
+                $groupInfo = Get-MgGroup -Filter "DisplayName eq '$_'"
+                $idValue = $_
+                if (-not [System.String]::IsNullOrEmpty($groupInfo))
+                {
+                    $idValue = $groupInfo.Id
+                }
+                $assignmentsArray += set-JSONstring -id $idValue -type 'Assignments'
             }
         }
         $configstring += ( 'Assignments' + ":`r`n" + ($PSBoundParameters.Assignments | Out-String) + "`r`n" )
@@ -721,7 +736,7 @@ function Set-TargetResource
         $setParams.add('AndroidManagedAppProtectionId', $currentPolicy.id)
         Update-MgBetaDeviceAppManagementAndroidManagedAppProtection @setParams
 
-        Write-Verbose -Message 'Setting Group Assignments...'
+        Write-Verbose -Message "Setting Group Assignments with values:`r`n$(ConvertTo-Json $assignmentsArray -Depth 10)"
         Set-MgBetaDeviceAppManagementTargetedManagedAppConfiguration -TargetedManagedAppConfigurationId $setParams.AndroidManagedAppProtectionId -Assignments $assignmentsArray
 
     }
@@ -1027,7 +1042,20 @@ function Test-TargetResource
     # handle complex parameters - manually for now
     if ($PSBoundParameters.keys -contains 'Assignments' )
     {
-        $targetvalues.add('Assignments', $psboundparameters.Assignments)
+        $assignmentsValue = @()
+        foreach ($assignment in $Assignments)
+        {
+            $groupInfo = Get-MgGroup -GroupId $assignment -ErrorAction SilentlyContinue
+            if ($null -ne $groupInfo)
+            {
+                $assignmentsValue += $groupInfo.DisplayName
+            }
+            else
+            {
+                $assignmentsValue += $assignment
+            }
+        }
+        $targetvalues.add('Assignments', $assignmentsValue)
     }
 
     Write-Verbose -Message 'Starting Exluded Groups Check'