+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneApplicationControlPolicyWindows10/MSFT_IntuneApplicationControlPolicyWindows10.psm1
@@ -91,7 +91,7 @@ function Get-TargetResource
             #Retrieve policy general settings
             if (-not [System.String]::IsNullOrEmpty($Id))
             {
-                $policy = Get-MgBetaDeviceManagementIntent -DeviceManagementIntentId $Id -ErrorAction Stop
+                $policy = Get-MgBetaDeviceManagementIntent -DeviceManagementIntentId $Id -ErrorAction SilentlyContinue
             }
 
             if ($null -eq $policy)