+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADRoleEligibilityScheduleRequest.Tests.ps1
@@ -48,14 +48,16 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             Mock -CommandName Get-MgBetaRoleManagementDirectoryRoleDefinition -MockWith {
                 return @{
-                    DisplayName = 'Teams Communications Administrator'
-                    Id          = '12345'
+                    DisplayName      = 'Teams Communications Administrator'
+                    Id               = '12345'
+                    DirectoryScopeId = '/'
                 }
             }
             Mock -CommandName Get-MgBetaRoleManagementDirectoryRoleEligibilitySchedule -MockWith {
                 return @{
                     Id          = '12345-12345-12345-12345-12345'
                     RoleDefinitionId = "12345"
+                    DirectoryScopeId = '/'
                 }
             }
 
@@ -87,6 +89,11 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 }
 
                 Mock -CommandName Get-MgBetaRoleManagementDirectoryRoleEligibilityScheduleRequest -MockWith {
+                    return $null
+                }
+                
+                
+                Mock -CommandName Get-MgBetaRoleManagementDirectoryRoleEligibilitySchedule -MockWith {
                     return $null
                 }
             }
@@ -263,7 +270,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should call the Set to Update the instance' {
                 Set-TargetResource @testParams
-                Should -Invoke -CommandName Get-MgBetaRoleManagementDirectoryRoleEligibilityScheduleRequest -Exactly 1
+                Should -Invoke -CommandName New-MgBetaRoleManagementDirectoryRoleEligibilityScheduleRequest -Exactly 1
             }
         }
         Context -Name 'ReverseDSC Tests' -Fixture {